Making server runnable

+++ Terraria/ID/SetFactory.cs
@@ -154,7 +154,7 @@ namespace Terraria.ID
 			}
 			for (int j = 0; j < inputs.Length; j += 2)
 			{
-				array[(int)((short)inputs[j])] = (T)((object)inputs[j + 1]);
+				array[(int)((short)(int)inputs[j])] = (T)((object)inputs[j + 1]);
 			}
 			return array;
 		}
